% $Id: sc00-SuperCurve-package.Rd 913 2014-08-05 22:51:12Z proebuck $

\name{SuperCurve-package}
\alias{SuperCurve-package}
\docType{package}
\title{Reverse phase protein lysate array analysis}
\description{
  A package for analyzing reverse phase protein lysate arrays (RPPA).
}
\details{
  \tabular{ll}{
    Package: \tab SuperCurve\cr
    Type:    \tab Package\cr
    Version: \tab \Sexpr[results=text]{packageDescription("SuperCurve")$Version}\cr
    Phase:   \tab \Sexpr[results=text]{packageDescription("SuperCurve")$Phase}\cr
    Date:    \tab \Sexpr[results=text]{packageDescription("SuperCurve")$Date}\cr
    License: \tab \Sexpr[results=text]{packageDescription("SuperCurve")$License}\cr
  }

  For a complete list of functions, use \code{library(help="SuperCurve")}.\cr
  For a high-level summary of the changes for each revision, use\cr
    \code{file.show(system.file("NEWS", package="SuperCurve"))}.
}
\author{
  Kevin R. Coombes \email{kcoombes@mdanderson.org},
  P. Roebuck \email{proebuck@mdanderson.org}
}
\keyword{package}

